クラスの責務を変更した。また、DisplayクラスをFileFormatterクラスに変更した
FileFormatterクラスは、ファイル情報を整形するようにした

LsCommandクラスを追加し、それぞれのクラスで整形された配列を表示するようにした

ls.rbでは、LsCommandクラスを実行するファイルにした
